Split tests from library code in bevy_landmass to match landmass.
